Day-15

+++ solutions/2. String/49. Binary-String.cpp
@@ -17,7 +17,7 @@
 class Solution {
     public:
     long binarySubstring(int n, string a){
-        int onesCount = 0;
+        long onesCount = 0;
         for(int i = 0; i < n; i++) {
             if(a[i] == '1') {
                 onesCount++;
